requirements.txt
# FastAPI and Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure SDK for Python
azure-identity==1.15.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4
azure-ai-contentsafety==1.0.0

# MongoDB and Motor (Async MongoDB Driver)
motor==3.3.2
pymongo==4.6.1
dnspython==2.4.2

# Azure Monitoring and Telemetry
azure-monitor-opentelemetry==1.2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Replicate API for Image Generation
replicate==0.23.1

# Authentication and JWT
PyJWT==2.8.0
cryptography==42.0.0
msal==1.26.0

# Logging and Monitoring
python-json-logger==2.0.7
structlog==24.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Image Processing
Pillow==10.2.0
pillow-heif==0.14.0

# Testing (optional, for tests/)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
